+++ demo/src/pages/TravelPreference.tsx
@@ -12,9 +12,12 @@ const TravelPreferenceForm = (i: {
   onSaved?: () => void;
 }) => {
 
+  const daysOfWeek = [ 'Monday', 'Tuesday', 'Wednesday', 'Thursday', 'Friday', 'Saturday', 'Sunday' ];
+  const [ isCommuter, setIsCommuter ] = useState<boolean>(true);
   const [ selectedResource, setSelectedResource ] = useState<string | undefined>(undefined);
   const [ travelPreference, setTravelPreference ] = useState<TravelPreference | undefined>(undefined);
   const [ modeOfTransportation, setModeOfTransportation ] = useState<string>('');
+  const [ selectedDaysOfWeek, setSelectedDaysOfWeek ] = useState<number[]>([]);
   // States the component can be in
   const [ state, setState ] = useState<'INITIAL' | 'LOADING' | 'LOADED' | 'ERROR'>('INITIAL');
 
@@ -24,17 +27,21 @@ const TravelPreferenceForm = (i: {
 
       try {
 
+        setIsCommuter(decodeIDToken(i.token).payload.webid !== import.meta.env.VITE_SUBJECT_WEBID);
+
         if(i.resource){
 
           const content = await retrieveData(i.resource, i.token, i.publicKey, i.privateKey);
           const parsedTravelPreference = parseTravelPreference(i.resource, content);
 
-          setTravelPreference(parsedTravelPreference as { modeOfTransportation: string });
+          setTravelPreference(parsedTravelPreference);
           setModeOfTransportation(parsedTravelPreference.modeOfTransportation);
+          setSelectedDaysOfWeek(parsedTravelPreference.daysOfWeek);
 
         } else {
 
           setModeOfTransportation('');
+          setSelectedDaysOfWeek([]);
 
         }
 
@@ -80,6 +87,7 @@ const TravelPreferenceForm = (i: {
       decodeIDToken(i.token).payload.webid,
       JSON.stringify({
         modeOfTransportation,
+        daysOfWeek: selectedDaysOfWeek,
       }),
       'application/json',
       i.token,
@@ -95,30 +103,62 @@ const TravelPreferenceForm = (i: {
 
   };
 
+  const onClickDayOfWeek = (index: number) => {
+
+    if(isCommuter){
+
+      setSelectedDaysOfWeek(
+        [
+          ...selectedDaysOfWeek.includes(index) ? [] : [ index ],
+          ...selectedDaysOfWeek.filter((dayOfWeekIndex) => dayOfWeekIndex !== index),
+        ]
+      );
+
+    }
+
+  };
+
   return (
     <>
       {state === 'ERROR' ? <p>Something went wrong</p> :
         <div className="w-full h-full flex flex-col rounded-lg bg-white">
           <h2 className="text-3xl p-4 border-b-2 border-slate-50 h-20">Travel preference</h2>
           <div className="flex flex-col content-between justify-between gap-2 text-sm p-4">
-            <input className="border border-emerald-500 px-3 py-2 rounded-md" placeholder="Mode of transportation" type="text" value={modeOfTransportation} onChange={(e) => setModeOfTransportation(e.target.value)}></input>
-            <div className="flex flex-row gap-4 justify-end">
-              <button
-                className="self-end hover:bg-emerald-600 bg-emerald-500 disabled:bg-slate-200 disabled:text-slate-400 disabled:cursor-default px-3 py-2 rounded-full cursor-pointer"
-                disabled={modeOfTransportation === '' || i.resource !== undefined}
-                onClick={
-                  () => {
-
-                    void (async () => {
-
-                      await onSave();
-
-                    })();
-
-                  }
-                }
-              >Save</button>
+            <input className="border border-emerald-500 px-3 py-2 rounded-md" disabled={!isCommuter} placeholder="Mode of transportation" type="text" value={modeOfTransportation} onChange={(e) => setModeOfTransportation(e.target.value)}></input>
+            <div className="flex flex-row gap-4">
+              {
+                // Show the days of the week.
+                daysOfWeek.map(
+                  (day, index) =>
+                    <div
+                      key={index}
+                      className={`rounded-md cursor-pointer border border-emerald-500 px-3 py-2 ${selectedDaysOfWeek.includes(index) ? 'bg-emerald-500' : 'bg-slate-100'}`}
+                      onClick={() => onClickDayOfWeek(index)}>
+                      {day}</div>
+                )
+              }
             </div>
+            {
+              // Only show save button when commuter is signed in
+              isCommuter ?
+                <div className="flex flex-row gap-4 justify-end">
+                  <button
+                    className="self-end hover:bg-emerald-600 bg-emerald-500 disabled:bg-slate-200 disabled:text-slate-400 disabled:cursor-default px-3 py-2 rounded-full cursor-pointer"
+                    disabled={modeOfTransportation === '' || i.resource !== undefined}
+                    onClick={
+                      () => {
+
+                        void (async () => {
+
+                          await onSave();
+
+                        })();
+
+                      }
+                    }
+                  >Save</button>
+                </div> : ''
+            }
           </div>
         </div>
       }