+++ sdk/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @useid/movejs
 
+## 0.13.0
+
+### Minor Changes
+
+- Travel preferences form and basic itinerary
+
 ## 0.12.0
 
 ### Minor Changes